# schemas/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "SWEN Bank API",
    "description": "API Documentation For SWEN Bank",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3030",
  "basePath": "/",
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/Users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new user account. Must be login as Admin or Employee role to access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          },
          {
            "name": "Request Body",
            "in": "body",
            "description": "Provide the user information to create a new user account",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userSchema"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User account is created",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "403": {
            "description": "Do not have permission to access the API",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "409": {
            "description": "User already exist",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update user account by id. Must be login as User role to access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          },
          {
            "name": "Request Body",
            "in": "body",
            "description": "Provide the user information to be updated",
            "required": true,
            "schema": {
              "properties": {
                "username": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "phone": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Update actor successfully",
            "schema": {
              "properties": {
                "username": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "phone": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
    "/Users/Login": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Login to Internet Banking web application",
        "parameters": [
          {
            "name": "Request Body",
            "in": "body",
            "description": "Provide the information to login",
            "required": true,
            "schema": {
              "$ref": "#/definitions/loginToWeb"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successfully",
            "schema": {
              "$ref": "#/definitions/loginResponse"
            }
          },
          "401": {
            "description": "Wrong username or password",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
    "/Users/ResetPassword": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Reset password of a user account",
        "parameters": [
          {
            "name": "Request Body",
            "in": "body",
            "description": "Provide the username of the account need to reset password",
            "required": true,
            "schema": {
              "properties": {
                "username": {
                  "type": "string"
                }
              },
              "required": [
                "username"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful message, OTP code has been sent",
            "schema": {
              "$ref": "#/definitions/successResponse"
            }
          },
          "401": {
            "description": "Invalid username",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
    "/Users/VerifyOTP": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Verify OTP code to reset password of a user account",
        "parameters": [
          {
            "name": "Request Body",
            "in": "body",
            "description": "Provide the OTP code and new password of the account",
            "required": true,
            "schema": {
              "properties": {
                "otp": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": [
                "otp",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully reset password",
            "schema": {
              "$ref": "#/definitions/successResponse"
            }
          },
          "401": {
            "description": "Invalid OTP code",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
    "/Users/Employees": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get all employees. Only admin can access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "List of employees",
            "schema": {
              "properties": {
                "id": {
                  "type": "integer"
                },
                "username": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "phone": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "role_id": {
                  "type": "integer"
                }
              }
            }
          },
          "403": {
            "description": "Do not have permission to access the API",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
    "/Users/Employees/{empployeeID}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete an empployee. Only admin can access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          },
          {
            "name": "empployeeID",
            "in": "path",
            "required": true,
            "description": "ID of an employee",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted employee",
            "schema": {
              "$ref": "#/definitions/successResponse"
            }
          },
          "403": {
            "description": "Do not have permission to access the API",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "user": {
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "role_id": {
          "type": "integer"
        }
      }
    },
...
    "loginResponse": {
      "properties": {
        "message": {
          "type": "string"
        },
        "token": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        }
      }
    },
...
    "userSchema": {
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "role_id": {
          "type": "integer"
        }
      },
      "required": [
        "username",
        "password",
        "email",
        "phone",
        "name"
      ]
    },
...
    "loginToWeb": {
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "username",
        "password"
      ]
    },
...
    "errorResponse": {
      "properties": {
        "message": {
          "type": "string",
          "description": "The error message"
        }
      }
    },
    "successResponse": {
      "properties": {
        "message": {
          "type": "string",
          "description": "The success message"
        }
      }
    }
  }
}